+++ py/symbolic/exceptions.py
@@ -1,27 +1,35 @@
+from __future__ import annotations
+from typing import TYPE_CHECKING
+
 from symbolic._lowlevel import lib
 
 
 __all__ = ["SymbolicError"]
-exceptions_by_code = {}
+exceptions_by_code: dict[int, type[SymbolicError]] = {}
 
 
 class SymbolicError(Exception):
-    code = None
+    if TYPE_CHECKING:
+        code: int
+    else:
+        code = None
 
     def __init__(self, msg):
         Exception.__init__(self)
         self.message = msg
         self.rust_info = None
 
-    def __str__(self):
+    def __str__(self) -> str:
         rv = self.message
         if self.rust_info is not None:
             return f"{rv}\n\n{self.rust_info}"
         return rv
 
 
-def _make_error(error_name, base=SymbolicError, code=None):
-    class Exc(base):
+def _make_error(
+    error_name: str, base: type[SymbolicError] = SymbolicError, code: int | None = None
+) -> type[SymbolicError]:
+    class Exc(base):  # type: ignore[misc,valid-type]
         pass
 
     Exc.__name__ = Exc.__qualname__ = error_name
@@ -32,7 +40,7 @@ def _make_error(error_name, base=SymbolicError, code=None):
     return Exc
 
 
-def _get_error_base(error_name):
+def _get_error_base(error_name: str) -> type[SymbolicError]:
     pieces = error_name.split("Error", 1)
     if len(pieces) == 2 and pieces[0] and pieces[1]:
         base_error_name = pieces[0] + "Error"
@@ -43,7 +51,7 @@ def _get_error_base(error_name):
     return SymbolicError
 
 
-def _make_exceptions():
+def _make_exceptions() -> None:
     for attr in dir(lib):
         if not attr.startswith("SYMBOLIC_ERROR_CODE_"):
             continue
@@ -55,3 +63,8 @@ def _make_exceptions():
 
 
 _make_exceptions()
+
+if TYPE_CHECKING:
+    # treat unknown attribute names as exception types
+    def __getattr__(name: str) -> type[SymbolicError]:
+        ...